% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_ump.R
\name{evaluate_ump}
\alias{evaluate_ump}
\title{Underdetermination diagnositc for Bayesian stable isotope mixing models}
\usage{
evaluate_ump(simm_outs, prob = 0.95)
}
\arguments{
\item{simm_outs}{A simm (and mcmc.list) object, whose posterior is beta.}

\item{prob}{A numeric scalar in the interval (0,1), giving highest posterior density (HPD) interval.}
}
\value{
A data.frame where each column represents:
\tabular{ll}{
\code{lower} \tab \code{:} lower bound of HPD interval of beta-dependent SIMM \cr
\code{upper} \tab \code{:} upper bound of HPD interval of beta-dependent SIMM \cr
\code{delta} \tab \code{:} the difference between lower and upper bounds \cr
}
}
\description{
\code{evaluate_ump} produces underdetermination diagnostics.
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

# Run MCMC for beta-dependent SIMM
out_H = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1000,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

evaluate_ump(out_H, prob = 0.95)
}
