% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixingspace.R
\name{mixingspace}
\alias{mixingspace}
\title{Plot isotopic mixing space}
\usage{
mixingspace(
  mixture,
  sources,
  correct = NULL,
  axis = 1:2,
  source_names = NULL,
  element_names = NULL
)
}
\arguments{
\item{mixture}{A matrix with each target individual as a separate row and each isotope as
a separate column.}

\item{sources}{A matrix containing the mean and standard deviations of the source values for
each of the isotopes.}

\item{correct}{A matrix containing the mean and standard deviations of the fractional correction
values for each of the isotopes.}

\item{axis}{The element axes used as x and y in mixing space.}

\item{source_names}{The source names.}

\item{element_names}{The element names.}
}
\value{
A ggplot output.
}
\description{
\code{mixingspace} draws isotopic mixing space.
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

mixingspace(
    mixture, sources,
    source_names  = paste("Source", 1:nrow(sources)),
    element_names = c("d13C", "d15N"))
}
