% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Correlation plot}
\usage{
plot_corr(simm_outs)
}
\arguments{
\item{simm_outs}{A simm (and mcmc.list) object.}
}
\value{
A ggplot output.
}
\description{
\code{plot_corr} draws probability density plots (diagonal), heat maps (lower) and
correlation values (upper).
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

# Run MCMC for ordinary SIMM
out_L = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

# Correlation plot
plot_corr(out_L[,1:nrow(sources)])
}
