% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_post.R
\name{plot_post}
\alias{plot_post}
\title{Plot posterior densities of SIMMs}
\usage{
plot_post(
  simm_outs,
  simm_outs_beta = NULL,
  prob = 0.95,
  type = c("source", "raw")
)
}
\arguments{
\item{simm_outs}{A simm (and mcmc.list) object, whose posterior is normal.}

\item{simm_outs_beta}{A simm (and mcmc.list) object, whose posterior is beta.}

\item{prob}{A numeric scalar in the interval (0,1), giving highest posterior density interval.}

\item{type}{The output type. The default is "source", where only source contirbutions are used
in posterior density plots.}
}
\value{
A ggplot output.
}
\description{
\code{plot_post} draws posterior densities from ordinary and beta-dependent SIMMs.
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

# Run MCMC for ordinary SIMM
out_L = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

# Run MCMC for beta-dependent SIMM
out_H = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1000,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

# Posterior density plot
plot_post(out_L, out_H, type = "source")
}
