% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siarbeta.R
\name{siarbeta}
\alias{siarbeta}
\title{Beta-dependent stable isotope mixing model}
\usage{
siarbeta(
  mixture,
  sources,
  correct = 0,
  concdep = 1,
  alpha = 1,
  beta = 1,
  error = c("stock", "variance", "residual", "parnell"),
  chains = 3,
  iters = 5000,
  burns = 1000,
  thins = 4,
  silent = FALSE,
  source_names = NULL
)
}
\arguments{
\item{mixture}{A matrix with each target individual as a separate row and each isotope as a separate column.}

\item{sources}{A matrix containing the mean and standard deviations of the source values for
each of the isotopes.}

\item{correct}{A matrix containing the mean and standard deviations of the fractional correction
values for each of the isotopes.}

\item{concdep}{A matrix containing the mean and standard deviations of the concentration dependence
values for each of the isotopes. Note that this function does not use the standard deviations.}

\item{alpha}{A scalar or vector containing the Dirichlet prior parameters.}

\item{beta}{The beta in beta-dependent posterior probability. The default is 1 (ordinary SIAR).}

\item{error}{The error structure. The default is "stock".}

\item{chains}{The number of chains. The default is 3.}

\item{iters}{The number of iterations to run. The default is 5000.}

\item{burns}{The size of the burn-in. The default is 1000.}

\item{thins}{The amount of thinning of the iterations. The default is 4.}

\item{silent}{prevents messages from being printed to the R console. The default is False.}

\item{source_names}{The source names.}
}
\value{
A mcmc.list output.
}
\description{
Applications of beta-dependent stable isotope mixing model.
\code{siarbeta} diagnoses the uncertainty due to underdetermination by beta-posteriors.
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

# Run MCMC for ordinary SIMM
out_L = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

# Run MCMC for beta-dependent SIMM
out_H = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1000,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)
}
