% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simm.R
\name{summary.simm}
\alias{summary.simm}
\title{Summary statisics for Bayesian stable isotope mixing models}
\usage{
\method{summary}{simm}(object, quantiles = c(0.025, 0.5, 0.975), ...)
}
\arguments{
\item{object}{A simm (and mcmc.list) object, which hasmore than one chain.}

\item{quantiles}{A vector of quantiles to evaluate credible intervals.}

\item{...}{Further arguments.}
}
\value{
A data.frame where each row represents summary statistic of SIMM outputs:
\tabular{ll}{
\code{Mean}    \tab \code{:} posterior mean \cr
\code{SD}      \tab \code{:} posterior standard deviation \cr
\code{2.5}     \tab \code{:} posterior 2.5 percent quantile \cr
\code{50}      \tab \code{:} posterior median \cr
\code{97.5}    \tab \code{:} posterior 97.5 percent quantile \cr
\code{rhat}    \tab \code{:} Gelman-Rubin convergence diagnostic \cr
\code{mean_r2} \tab \code{:} Mean determination coefficient \cr
}
}
\description{
\code{summary.simm} produces summary statistics (mean, SD, quantiles, Gelman diagnostic
and mean determination coefficinet).
}
\examples{
mixture = matrix(0, nrow = 1, ncol = 2)
sources = matrix(c(
    -1, 0.25,  1, 0.25,
    -1, 0.25, -1, 0.25,
     1, 0.25,  1, 0.25,
     1, 0.25, -1, 0.25), nrow = 4, byrow = TRUE)

# Run MCMC for ordinary SIMM
out_L = siarbeta(
    mixture, sources, correct = 0, concdep = 1, alpha = 1, beta = 1,
    error = "variance", chains = 3, iters = 10000, burns = 5000, thins = 4)

# Summary statistcs
summary.simm(out_L)
}
